Changes:   - Place scripts in the layout page to respect canonical tag placement   - Creat .js scripts for each page to better organize .cshtml files (breakdown, user, detailed)   - RefactoreCategoryController to have authentication and option of deletion to category   - Refactor ICategoryService and its implementing classes to be responsible for only handling category creation and deletion (expense assignment is responsability for ExpenseService)   - Detailed.cshtml to have alert on successful calendar adding of expenses   - New css for new elements in expenses.css   - Expense controller to redirect to current view on expense adding   - ExpenseService to be able to dissassociate expenses and categories   - HomeController to support logout   - IUserService and its implementing classes to have a new method for limit editing Add:   - Breakdown, Detailed and User javascript files to store each view necessary js   - edit.svg, an icon symbolizing something is editable   - User controller, view and page

+++ PersonalExpenses/Services/CategoryService.cs
@@ -13,63 +13,28 @@ public class CategoryService : ICategoryService
     {
         _context = context;
     }
-    public Task<Category> CreateCategory(string name, string color)
+    public async Task<Category> CreateCategory(string name, string color, int userId)
     {
-        Category category = new Category { Name = name, Color = color };
-        _context.Categories.Add(category);
-        return Task.FromResult(category);
+        Category category = new Category { Name = name, Color = color, UserId = userId };
+        await _context.Categories.AddAsync(category);
+        await _context.SaveChangesAsync();
+        return category;
     }
 
-    public Task<bool> DeleteCategory(int categoryId)
+    public async Task<bool> DeleteCategory(int categoryId, int userId)
     {
-        Category? category = _context.Categories.Find(categoryId);
-        if (category == null)
+        Category? category = await _context.Categories.FindAsync(categoryId);
+        if (category == null || category.UserId != userId)
         {
-            return Task.FromResult(false);
+            return false;
         }
         _context.Categories.Remove(category);
-        _context.SaveChanges();
-        return Task.FromResult(true);
-    }
-
-    public async Task<Category> GetCategory(int categoryId)
-    {
-        return await _context.Categories.FindAsync(categoryId) ?? throw new Exception("Category not found");
-    }
-
-    public Task<bool> AddExpenseToCategory(int categoryId, int expenseId)
-    {
-        Expense? expense = _context.Expenses.Find(expenseId);
-        if (expense == null)
-        {
-            return Task.FromResult(false);
-        }
-        Category? category = _context.Categories.Find(categoryId);
-        if (category == null || expense.CategoryId != 0)
-        {
-            return Task.FromResult(false);
-        }
-        expense.CategoryId = categoryId;
-        _context.SaveChanges();
-        return Task.FromResult(true);
+        await _context.SaveChangesAsync();
+        return true;
     }
 
     public async Task<List<Category>> GetUserCategories(int userId)
     {
         return await _context.Categories.Where(c => c.UserId == userId).ToListAsync();
-    }
-
-    public Task<bool> RemoveExpenseFromCategory(int categoryId, int expenseId)
-    {
-        Expense? expense = _context.Expenses.Find(expenseId);
-        if (expense == null) return Task.FromResult(false);
-        
-        Category? category = _context.Categories.Find(categoryId);
-        if (category == null) return Task.FromResult(false);
-        
-        if(!_context.Expenses.Any(e => e.Id == expenseId)) return Task.FromResult(false);
-        expense.CategoryId = 0;
-        _context.SaveChanges();
-        return Task.FromResult(true);
     }
 }
